changed submissions.score type to tinyint

+++ app/database.py
@@ -29,7 +29,7 @@ class Database:
                 "id"          BIGINT      PRIMARY KEY,
                 "user_id"     BIGINT      NOT NULL,
                 "problem_id"  BIGINT      NOT NULL,
-                "score"       REAL        NOT NULL,
+                "score"       TINYINT     NOT NULL,
                 UNIQUE ("user_id", "problem_id") ON CONFLICT REPLACE
             )
         ''')
